+++ docs/_includes/NEWS.md
@@ -34,7 +34,7 @@ _Not yet released_
 ### NTP support
 
 - SatPulse now supports the NTP SHM protocol in addition to the chrony refclock SOCK protocol for sending time information to an NTP server. `satpulse.toml` has a new `[ntp.shm]` table for configuring this. (#300)
-- The `[serial]` table in `satpulse.toml` has a new `pps.line` key that lets `satpulsed` detect receiver PPS edges on a serial modem-control input and use them for more precise refclock samples on systems without a PTP hardware clock. (#402)
+- The `[serial]` table in `satpulse.toml` has a new `pps.pin` key that lets `satpulsed` detect receiver PPS edges on a serial modem-control input and use them for more precise refclock samples on systems without a PTP hardware clock. (#402)
 
 ### GPS high-level configuration